requirements.txt
google-generativeai
python-dotenv
pyautogui
Pillow
opencv-python
numpy
pytesseract
psutil
pywin32